refactor: clean up imports and remove deprecated useMobile hook
