+++ src/components/menu-editor/menu-publish.tsx
@@ -50,10 +50,10 @@ import {
   updateMenuStatus
 } from "@/server/actions/menu/mutations"
 import type { getMenuById } from "@/server/actions/menu/queries"
+import useLocalStorage from "@/hooks/use-local-storage"
 import { colorThemeAtom, fontThemeAtom, tourModeAtom } from "@/lib/atoms"
 import exportAsImage from "@/lib/export-as-image"
 import { MenuStatus } from "@/lib/types"
-import useLocalStorage from "@/lib/use-local-storage"
 import { getBaseUrl } from "@/lib/utils"
 
 export default function MenuPublish({
