--- src/lib/use-mobile.tsx
@@ -1,16 +1,0 @@
-import { useEffect, useState } from "react"
-
-export function useMobile(breakpoint = 640) {
-  const [isMobile, setIsMobile] = useState(false)
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth <= breakpoint)
-    }
-    window.addEventListener("resize", handleResize)
-    handleResize()
-    return () => {
-      window.removeEventListener("resize", handleResize)
-    }
-  }, []) // eslint-disable-line react-hooks/exhaustive-deps
-  return isMobile
-}
